[test][theme] extract color definition assertion test helper
