+++ test/theme.go
@@ -3,12 +3,64 @@ package test
 import (
 	"fmt"
 	"image/color"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/theme"
 )
 
 var defaultTheme fyne.Theme
+
+// Try to keep these in sync with the existing color names at theme/color.go.
+var knownColorNames = []fyne.ThemeColorName{
+	theme.ColorNameBackground,
+	theme.ColorNameButton,
+	theme.ColorNameDisabled,
+	theme.ColorNameDisabledButton,
+	theme.ColorNameError,
+	theme.ColorNameFocus,
+	theme.ColorNameForeground,
+	theme.ColorNameForegroundOnError,
+	theme.ColorNameForegroundOnPrimary,
+	theme.ColorNameForegroundOnSuccess,
+	theme.ColorNameForegroundOnWarning,
+	theme.ColorNameHeaderBackground,
+	theme.ColorNameHover,
+	theme.ColorNameHyperlink,
+	theme.ColorNameInputBackground,
+	theme.ColorNameInputBorder,
+	theme.ColorNameMenuBackground,
+	theme.ColorNameOverlayBackground,
+	theme.ColorNamePlaceHolder,
+	theme.ColorNamePressed,
+	theme.ColorNamePrimary,
+	theme.ColorNameScrollBar,
+	theme.ColorNameSelection,
+	theme.ColorNameSeparator,
+	theme.ColorNameShadow,
+	theme.ColorNameSuccess,
+	theme.ColorNameWarning,
+}
+
+// Try to keep this in sync with the existing variants at theme/theme.go
+var knownVariants = []fyne.ThemeVariant{
+	theme.VariantDark,
+	theme.VariantLight,
+}
+
+// AssertAllColorNamesDefined asserts that all known color names are defined for the given theme.
+func AssertAllColorNamesDefined(t *testing.T, th fyne.Theme, themeName string) {
+	for _, variant := range knownVariants {
+		for _, cn := range knownColorNames {
+			assert.NotNil(t, th.Color(cn, variant), "undefined color %s variant %d in theme %s", cn, variant, themeName)
+			// Transparent is used by the default theme as fallback for unknown color names.
+			// Built-in color names should have well-defined non-transparent values.
+			assert.NotEqual(t, color.Transparent, th.Color(cn, variant), "undefined color %s variant %d in theme %s", cn, variant, themeName)
+		}
+	}
+}
 
 // NewTheme returns a new test theme using quiet ugly colors.
 func NewTheme() fyne.Theme {
