+++ theme/color.go
@@ -7,7 +7,7 @@ import (
 	internaltheme "fyne.io/fyne/v2/internal/theme"
 )
 
-// Keep in mind to add new constants to the tests at theme/theme_test.go as well as test/theme_test.go.
+// Keep in mind to add new constants to the tests at test/theme.go.
 const (
 	// ColorRed is the red primary color name.
 	//
